requirements.txt
aiogram==3.3.0
tortoise-orm==0.20.0
aiosqlite==0.19.0
python-dotenv==1.0.0
APScheduler==3.10.4
pytz==2023.3
pandas==2.1.4
validators==0.22.0



